Reorganized triggers when OnClientIsReady is fired.
